add leaderboard preview in stats page
