+++ src/components/stats/stats-page.tsx
@@ -1,6 +1,7 @@
 import Container from '@/components/container/container';
 import SectionTitle from '@/components/section-title/section-title';
 import JobsRevenueStats from '@/components/stats/jobs-revenue-stats';
+import LeaderboardPreview from '@/components/stats/leaderboard-preview';
 import NodeSpecStats from '@/components/stats/system-stats';
 import TopGpuModels from '@/components/stats/top-gpu-models';
 import TopNodes from '@/components/stats/top-nodes';
@@ -16,6 +17,7 @@ const StatsPage = () => {
       />
       <div className="pageContentWrapper">
         <JobsRevenueStats />
+        <LeaderboardPreview />
         <TopGpuModels />
         <NodeSpecStats />
         <TopNodes />
